backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2
nltk==3.9.1
spacy==3.8.2
transformers==4.45.2
torch==2.4.1
opencv-python==4.10.0.84
Pillow==10.4.0
fpdf2==2.7.9
numpy==2.1.2
scikit-learn==1.5.2
httpx==0.27.2
beautifulsoup4==4.12.3
